Refactored audio search forms into their own components.
