+++ AdminWebsite/AdminWebsite/ClientApp/src/app/get-audio-file/get-audio-file-vh/get-audio-file-vh.component.ts
@@ -1,0 +1,75 @@
+import { Component, OnInit } from '@angular/core';
+import { FormGroup, FormBuilder } from '@angular/forms';
+import { HearingAudioSearchModel } from 'src/app/common/model/hearing-audio-search-model';
+import { AudioLinkService } from 'src/app/services/audio-link-service';
+import { Logger } from 'src/app/services/logger';
+
+@Component({
+  selector: 'app-get-audio-file-vh',
+  templateUrl: './get-audio-file-vh.component.html',
+  styleUrls: ['./get-audio-file-vh.component.css']
+})
+export class GetAudioFileVhComponent implements OnInit {
+    private readonly loggerPrefix = '[GetAudioFile] -';
+    getVhAudioFileForm: FormGroup;
+    hasSearched: boolean;
+    results: HearingAudioSearchModel[] = [];
+    today = new Date();
+
+    constructor(private fb: FormBuilder, private audioLinkService: AudioLinkService, private logger: Logger) {
+        this.hasSearched = false;
+    }
+
+    async ngOnInit(): Promise<void> {
+        this.logger.debug(`${this.loggerPrefix} Landed on get audio file`);
+
+        this.getVhAudioFileForm = this.fb.group({
+            caseNumber: [null],
+            vhDate: [null]
+        });
+    }
+
+    get caseNumber() {
+        return this.getVhAudioFileForm.get('caseNumber');
+    }
+
+    get vhDate() {
+        return this.getVhAudioFileForm.get('vhDate');
+    }
+
+    get vhSearchCriteriaSet(): boolean {
+        return this.caseNumber.value || this.vhDate.value;
+    }
+
+    get vhDateInvalid() {
+        const todayDate = new Date(new Date());
+        return new Date(this.vhDate.value) > todayDate && (this.vhDate.dirty || this.vhDate.touched);
+    }
+
+    async search() {
+        this.logger.debug(`${this.loggerPrefix} Attempting to search for audio recording`);
+        if (this.vhSearchCriteriaSet) {
+            this.hasSearched = false;
+
+            const date: Date = this.vhDate.value ? new Date(this.vhDate.value) : undefined;
+            const caseNumber: string = this.caseNumber.value ? this.caseNumber.value : undefined;
+            this.results = await this.getResults(caseNumber, date);
+
+            this.hasSearched = true;
+        }
+    }
+
+    async getResults(caseNumber: string, date?: Date): Promise<HearingAudioSearchModel[]> {
+        this.logger.debug(`${this.loggerPrefix} Getting results by case number/date`, { caseNumber, date });
+        const response = await this.audioLinkService.searchForHearingsByCaseNumberOrDate(caseNumber, date);
+
+        if (response === null) {
+            this.logger.warn(`${this.loggerPrefix} No results`, { caseNumber, date });
+            return [];
+        }
+        this.logger.debug(`${this.loggerPrefix} Mapping results`, { caseNumber, date });
+        return response.map(x => {
+            return new HearingAudioSearchModel(x);
+        });
+    }
+}
