+++ AdminWebsite/AdminWebsite/ClientApp/src/app/get-audio-file/get-audio-file.component.html
@@ -16,7 +16,6 @@
             type="radio"
             [value]="'vhFile'"
             data-aria-controls="conditional-how-contacted-conditional"
-            (click)="searchChoiceClick()"
           />
           <label class="govuk-label govuk-radios__label" for="search-choice-vhfile"> Search for a video hearing audio file </label>
         </div>
@@ -29,91 +28,18 @@
             type="radio"
             [value]="'cvpFile'"
             data-aria-controls="conditional-how-contacted"
-            (click)="searchChoiceClick()"
           />
           <label class="govuk-label govuk-radios__label" for="search-choice-cvpfile"> Search for a CVP audio file </label>
         </div>
       </div>
     </div>
-    <div class="govuk-!-margin-top-9" *ngIf="searchChoice === 'vhFile'">
-      <label class="govuk-label govuk-!-width-one-half" for="caseNumber"> Enter a case number </label>
-      <input id="caseNumber" class="govuk-input govuk-!-width-one-quarter" type="text" formControlName="caseNumber" required />
-      <div class="govuk-!-margin-top-3">
-        <div [ngClass]="vhDateInvalid ? 'govuk-form-group--error' : 'govuk-form-group'">
-          <label class="govuk-label govuk-!-width-one-quarter" for="vhDate"> Date </label>
-          <input
-            id="vhDate"
-            class="govuk-input govuk-!-width-one-quarter"
-            type="date"
-            formControlName="vhDate"
-            required
-            placeholder="e.g. 30/08/2020"
-            max="{{ today | date: 'yyyy-MM-dd' }}"
-          />
-          <div *ngIf="vhDateInvalid" class="alert alert-danger">
-            <span id="vhDate-error" class="govuk-error-message"> Please enter a date in the past </span>
-          </div>
-        </div>
-      </div>
-      <button
-        class="govuk-button govuk-!-margin-top-3"
-        data-module="govuk-button"
-        id="submit"
-        (click)="search()"
-        [disabled]="!vhSearchCriteriaSet"
-      >
-        Search
-      </button>
-    </div>
-
-    <div class="govuk-!-margin-top-9" *ngIf="searchChoice === 'cvpFile'">
-      <div [ngClass]="hearingDateInvalid ? 'govuk-form-group--error' : 'govuk-form-group'">
-        <label class="govuk-label govuk-!-width-one-quarter" for="hearingDate"> Date </label>
-        <input
-          id="hearingDate"
-          class="govuk-input govuk-!-width-one-quarter"
-          type="date"
-          formControlName="hearingDate"
-          required
-          placeholder="e.g. 30/08/2020"
-          max="{{ today | date: 'yyyy-MM-dd' }}"
-        />
-        <div *ngIf="hearingDateInvalid" class="alert alert-danger">
-          <span id="hearingDate-error" class="govuk-error-message"> Please enter a date in the past </span>
-        </div>
-      </div>
-    </div>
-
-    <div class="govuk-!-margin-top-3" *ngIf="searchChoice === 'cvpFile'">
-      <div [ngClass]="cloudroomNameInvalid ? 'govuk-form-group--error' : 'govuk-form-group'">
-        <label class="govuk-label govuk-!-width-one-quarter" for="cloudroomName"> Cloudroom number </label>
-        <input id="cloudroomName" class="govuk-input govuk-!-width-one-half" type="text" formControlName="cloudroomName" />
-        <div *ngIf="cloudroomNameInvalid" class="alert alert-danger">
-          <span id="hearingDate-error" class="govuk-error-message"> Enter the number of the cloudroom </span>
-        </div>
-      </div>
-
-      <div class="govuk-!-margin-top-3">
-        <label class="govuk-label govuk-!-width-two-thirds" for="caseReference"> Case reference </label>
-        <input id="caseReference" class="govuk-input govuk-!-width-one-half" type="text" formControlName="caseReference" />
-      </div>
-      <button
-        class="govuk-button govuk-!-margin-top-3"
-        data-module="govuk-button"
-        id="submitCvp"
-        (click)="searchCVP()"
-        [disabled]="cvpRequestInvalid"
-      >
-        Search
-      </button>
-    </div>
   </div>
 </form>
 
-<div class="govuk-body" *ngIf="hasSearched">
-  <app-hearing-search-results [results]="results"></app-hearing-search-results>
+<div *ngIf="searchChoice === 'vhFile'">
+    <app-get-audio-file-vh></app-get-audio-file-vh>
 </div>
 
-<div class="govuk-body" *ngIf="hasCvpSearched">
-  <app-cvp-audio-file-search-result [results]="cvpResults"></app-cvp-audio-file-search-result>
+<div *ngIf="searchChoice === 'cvpFile'">
+    <app-get-audio-file-cvp></app-get-audio-file-cvp>
 </div>
